Refactored base Prop types; split between consumable and boundable to a registry
