+++ props-core/src/main/java/sh/props/custom/DoubleProp.java
@@ -28,7 +28,6 @@ package sh.props.custom;
 import sh.props.CustomProp;
 import sh.props.annotations.Nullable;
 import sh.props.converters.DoubleConverter;
-import sh.props.interfaces.Prop;
 
 /** Convenience implementation that constructs a {@link CustomProp} of the underlying type. */
 public class DoubleProp extends CustomProp<Double> implements DoubleConverter {
@@ -49,7 +48,8 @@ public class DoubleProp extends CustomProp<Double> implements DoubleConverter {
    * @param key the Prop's key
    * @param defaultValue a default value, or null if one doesn't exist
    * @param description a generic description used to explain what the prop is for
-   * @param isRequired true if the Prop must have a value when {@link Prop#get()} is called
+   * @param isRequired true if the Prop must have a value when {@link sh.props.AbstractProp#get()}
+   *     is called
    * @param isSecret true if the Prop represents a secret, in which case its value will be redacted
    *     when {@link Object#toString()} is called
    */
