+++ props-core/src/main/java/sh/props/custom/ListOfFloatsProp.java
@@ -29,7 +29,6 @@ import java.util.List;
 import sh.props.CustomProp;
 import sh.props.annotations.Nullable;
 import sh.props.converters.ListOfFloatConverter;
-import sh.props.interfaces.Prop;
 
 /** Convenience implementation that constructs a {@link CustomProp} of the underlying type. */
 public class ListOfFloatsProp extends CustomProp<List<Float>> implements ListOfFloatConverter {
@@ -50,7 +49,8 @@ public class ListOfFloatsProp extends CustomProp<List<Float>> implements ListOfF
    * @param key the Prop's key
    * @param defaultValue a default value, or null if one doesn't exist
    * @param description a generic description used to explain what the prop is for
-   * @param isRequired true if the Prop must have a value when {@link Prop#get()} is called
+   * @param isRequired true if the Prop must have a value when {@link sh.props.AbstractProp#get()}
+   *     is called
    * @param isSecret true if the Prop represents a secret, in which case its value will be redacted
    *     when {@link Object#toString()} is called
    */
